updated app to read effects.json and render effects correctly
