+++ server_ui/package.json
@@ -29,6 +29,7 @@
     "babel-preset-es2015": "^6.18.0",
     "babel-preset-react": "^6.16.0",
     "html-webpack-plugin": "^2.26.0",
+    "json-loader": "^0.5.4",
     "webpack": "^1.14.0",
     "webpack-dev-server": "^1.16.2"
   }
